refactor: Modal 폴더 생성
